move subcomponents of members modal from screens to components
